interfaces/apparmor,apparmor: move apparmor to a new package
There's some generic apparmor code that is not tied to our interface
code that would be nice to import from other places but this cannot be
done because of cyclic imports.

This patch adds a new package and moves the probe code there.
